+++ cli/test/update.ios.spec.ts
@@ -4,11 +4,9 @@ import {
   CORDOVA_PLUGIN_ID,
   MappedFS,
   makeAppDir,
-  makeConfig,
   run,
   installPlatform,
 } from './util';
-import { updateCommand } from '../src/tasks/update';
 
 describe.each([false, true])('Update: iOS (monoRepoLike: %p)', monoRepoLike => {
   let appDirObj: any;
@@ -24,8 +22,6 @@ describe.each([false, true])('Update: iOS (monoRepoLike: %p)', monoRepoLike => {
     await run(appDir, `init "${APP_NAME}" "${APP_ID}"`);
     await installPlatform(appDir, 'ios');
     await run(appDir, `add ios`);
-    // Redundant, because add does this, but called explicitly for thoroughness
-    await updateCommand(makeConfig(appDir), 'ios', false);
     FS = new MappedFS(appDir);
   });
 
@@ -33,13 +29,11 @@ describe.each([false, true])('Update: iOS (monoRepoLike: %p)', monoRepoLike => {
     //appDirObj.cleanupCallback();
   });
 
-  it('Should update', async () => {});
-
   it('Should install Cordova plugin JS', async () => {
     const cordovaPluginJSContent = await FS.read(
       'ios/App/public/cordova_plugins.js',
     );
-    let regex = new RegExp(CORDOVA_PLUGIN_ID);
+    const regex = new RegExp(CORDOVA_PLUGIN_ID);
     expect(regex.test(cordovaPluginJSContent)).toBe(true);
   });